# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HexTrack API",
    "description": "League of Legends stats with an AI Score.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/summoners/by-riot-id/{game_name}/{tag_line}": {
      "get": {
        "tags": [
          "summoners"
        ],
        "summary": "Profile by Riot ID (resolves unknown players through Riot)",
        "operationId": "get_summoner_profile",
        "parameters": [
          {
            "name": "game_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 32,
              "pattern": "^[^\\x00-\\x1f\\x7f]+$",
              "description": "Riot ID game name",
              "title": "Game Name"
            },
            "description": "Riot ID game name"
          },
          {
            "name": "tag_line",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 8,
              "pattern": "^[^\\x00-\\x1f\\x7f]+$",
              "description": "Riot ID tag line",
              "title": "Tag Line"
            },
            "description": "Riot ID tag line"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummonerProfile"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Riot API rate limited (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Riot API unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Riot API key missing/rejected, or AI model not loaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summoners/by-riot-id/{game_name}/{tag_line}/refresh": {
      "post": {
        "tags": [
          "summoners"
        ],
        "summary": "Update button: refresh rank and ingest new matches (rate limited per player)",
        "operationId": "refresh_summoner",
        "parameters": [
          {
            "name": "game_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 32,
              "pattern": "^[^\\x00-\\x1f\\x7f]+$",
              "description": "Riot ID game name",
              "title": "Game Name"
            },
            "description": "Riot ID game name"
          },
          {
            "name": "tag_line",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 8,
              "pattern": "^[^\\x00-\\x1f\\x7f]+$",
              "description": "Riot ID tag line",
              "title": "Tag Line"
            },
            "description": "Riot ID tag line"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Riot API rate limited (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Riot API unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Riot API key missing/rejected, or AI model not loaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summoners/{puuid}/matches": {
      "get": {
        "tags": [
          "summoners"
        ],
        "summary": "Match history, newest first (cursor paginated)",
        "operationId": "list_summoner_matches",
        "parameters": [
          {
            "name": "puuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[^\\x00-\\x1f\\x7f]+$",
              "title": "Puuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "queue",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "maximum": 2147483647,
                    "minimum": 0
                  },
                  "maxItems": 10
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by queue id; repeat the parameter to match any of several (e.g. queue=400&queue=430 for normals)",
              "title": "Queue"
            },
            "description": "Filter by queue id; repeat the parameter to match any of several (e.g. queue=400&queue=430 for normals)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchPage"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summoners/{puuid}/ranks": {
      "get": {
        "tags": [
          "summoners"
        ],
        "summary": "Rank snapshots for one queue, oldest first",
        "operationId": "get_rank_history",
        "parameters": [
          {
            "name": "puuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[^\\x00-\\x1f\\x7f]+$",
              "title": "Puuid"
            }
          },
          {
            "name": "queue",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "RANKED_SOLO_5x5",
                "RANKED_FLEX_SR"
              ],
              "type": "string",
              "default": "RANKED_SOLO_5x5",
              "title": "Queue"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankHistory"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChampionStat": {
        "properties": {
          "champion_id": {
            "type": "integer",
            "title": "Champion Id"
          },
          "champion_name": {
            "type": "string",
            "title": "Champion Name"
          },
          "games": {
            "type": "integer",
            "title": "Games"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "winrate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Winrate"
          },
          "kda": {
            "type": "number",
            "title": "Kda"
          },
          "avg_ai_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Ai Score"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "champion_id",
          "champion_name",
          "games",
          "wins",
          "winrate",
          "kda",
          "avg_ai_score"
        ],
        "title": "ChampionStat"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MatchSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "MatchPage"
      },
...
      "ProfileStats": {
        "properties": {
          "games": {
            "type": "integer",
            "title": "Games"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "losses": {
            "type": "integer",
            "title": "Losses"
          },
          "winrate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Winrate"
          },
          "avg_kills": {
            "type": "number",
            "title": "Avg Kills"
          },
          "avg_deaths": {
            "type": "number",
            "title": "Avg Deaths"
          },
          "avg_assists": {
            "type": "number",
            "title": "Avg Assists"
          },
          "kda": {
            "type": "number",
            "title": "Kda"
          },
          "avg_cs_per_min": {
            "type": "number",
            "title": "Avg Cs Per Min"
          },
          "avg_damage_per_min": {
            "type": "number",
            "title": "Avg Damage Per Min"
          },
          "avg_vision_per_min": {
            "type": "number",
            "title": "Avg Vision Per Min"
          },
          "avg_kill_participation": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Avg Kill Participation"
          },
          "avg_ai_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Ai Score"
          },
          "ai_scored_games": {
            "type": "integer",
            "title": "Ai Scored Games"
          },
          "avg_ai_role_percentile": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Ai Role Percentile"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "games",
          "wins",
          "losses",
          "winrate",
          "avg_kills",
          "avg_deaths",
          "avg_assists",
          "kda",
          "avg_cs_per_min",
          "avg_damage_per_min",
          "avg_vision_per_min",
          "avg_kill_participation",
          "avg_ai_score",
          "ai_scored_games",
          "avg_ai_role_percentile"
        ],
        "title": "ProfileStats",
        "description": "Season aggregates over ranked queues (420, 440), remakes excluded."
      },
...
      "RankEntry": {
        "properties": {
          "queue_type": {
            "type": "string",
            "enum": [
              "RANKED_SOLO_5x5",
              "RANKED_FLEX_SR"
            ],
            "title": "Queue Type"
          },
          "tier": {
            "type": "string",
            "enum": [
              "IRON",
              "BRONZE",
              "SILVER",
              "GOLD",
              "PLATINUM",
              "EMERALD",
              "DIAMOND",
              "MASTER",
              "GRANDMASTER",
              "CHALLENGER"
            ],
            "title": "Tier"
          },
          "rank": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "I",
                  "II",
                  "III",
                  "IV"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "lp": {
            "type": "integer",
            "title": "Lp"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "losses": {
            "type": "integer",
            "title": "Losses"
          },
          "winrate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Winrate"
          },
          "rank_value": {
            "type": "integer",
            "title": "Rank Value"
          },
          "taken_at": {
            "type": "string",
            "format": "date-time",
            "title": "Taken At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "queue_type",
          "tier",
          "rank",
          "lp",
          "wins",
          "losses",
          "winrate",
          "rank_value",
          "taken_at"
        ],
        "title": "RankEntry"
      },
      "RankHistory": {
        "properties": {
          "puuid": {
            "type": "string",
            "title": "Puuid"
          },
          "queue_type": {
            "type": "string",
            "enum": [
              "RANKED_SOLO_5x5",
              "RANKED_FLEX_SR"
            ],
            "title": "Queue Type"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/RankPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "puuid",
          "queue_type",
          "points"
        ],
        "title": "RankHistory"
      },
...
      "RefreshResult": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "cooldown",
              "partial",
              "unavailable"
            ],
            "title": "Status"
          },
          "new_matches": {
            "type": "integer",
            "title": "New Matches"
          },
          "pending": {
            "type": "integer",
            "title": "Pending"
          },
          "next_allowed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Next Allowed At"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "profile": {
            "$ref": "#/components/schemas/SummonerProfile"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "new_matches",
          "pending",
          "next_allowed_at",
          "message",
          "profile"
        ],
        "title": "RefreshResult"
      },
      "RoleStat": {
        "properties": {
          "position": {
            "type": "string",
            "enum": [
              "TOP",
              "JUNGLE",
              "MIDDLE",
              "BOTTOM",
              "UTILITY",
              "UNKNOWN"
            ],
            "title": "Position"
          },
          "games": {
            "type": "integer",
            "title": "Games"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "winrate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Winrate"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "position",
          "games",
          "wins",
          "winrate"
        ],
        "title": "RoleStat"
      },
...
      "SummonerProfile": {
        "properties": {
          "puuid": {
            "type": "string",
            "title": "Puuid"
          },
          "game_name": {
            "type": "string",
            "title": "Game Name"
          },
          "tag_line": {
            "type": "string",
            "title": "Tag Line"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "profile_icon_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Icon Id"
          },
          "summoner_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summoner Level"
          },
          "is_tracked": {
            "type": "boolean",
            "title": "Is Tracked"
          },
          "tracked_since": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracked Since"
          },
          "last_refreshed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Refreshed At"
          },
          "can_refresh_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Can Refresh At"
          },
          "solo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RankEntry"
              },
              {
                "type": "null"
              }
            ]
          },
          "flex": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RankEntry"
              },
              {
                "type": "null"
              }
            ]
          },
          "stats": {
            "$ref": "#/components/schemas/ProfileStats"
          },
          "top_champions": {
            "items": {
              "$ref": "#/components/schemas/ChampionStat"
            },
            "type": "array",
            "maxItems": 7,
            "title": "Top Champions"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleStat"
            },
            "type": "array",
            "title": "Roles"
          },
          "recent_form": {
            "items": {
              "type": "boolean"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Recent Form"
          },
          "main_champion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Main Champion"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "puuid",
          "game_name",
          "tag_line",
          "platform",
          "profile_icon_id",
          "summoner_level",
          "is_tracked",
          "tracked_since",
          "last_refreshed_at",
          "can_refresh_at",
          "solo",
          "flex",
          "stats",
          "top_champions",
          "roles",
          "recent_form",
          "main_champion"
        ],
        "title": "SummonerProfile"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "HEXTRACK_ADMIN_TOKEN",
        "scheme": "bearer"
      }
    }
  }
}